feat: refactor foreground demo
