+++ AGENTS.md
@@ -50,6 +50,7 @@ This file tracks Codex progress and upcoming tasks. Keep it chronological and ap
 - **36** – Updated `randomPaint.frag` to blend snapshot alpha with history so blurred edges accumulate correctly. Lint and build pass.
 - **37** – Switched `randomPaint.frag` to premultiplied alpha blending to remove dark edges. Lint and build pass.
 - **38** – Revised `randomPaint.frag` to use alpha compositing with decay, preventing color blowout and keeping edges smooth. Lint and build pass.
+- **39** – Refactored `ForegroundLayer` into `SvgMesh`, extracted `DraggableSvg` and `DemoControls` components, and updated README. Lint and build pass.
 
 ## Next Steps
